Refactored to extract an ArticlesLoader component
